+++ sql/api/src/main/scala/org/apache/spark/sql/ForeachWriter.scala
