+++ sql/api/src/main/scala/org/apache/spark/sql/streaming/StreamingQueryListener.scala
